requirements.txt
langchain==0.3.4 
langchain-community==0.3.3 
langchain-groq==0.2.0
pydub==0.25.1
elevenlabs==1.11.0
python-dotenv==1.0.1
fastapi==0.114.0
uvicorn==0.30.6